refactored getters
