+++ lib/worker.js
@@ -1,22 +1,28 @@
+/*eslint-env node */
+'use strict';
 
-
-module.exports = {
-  setWorkerName: function(){
+module.exports = function(Queue){
+  
+  Queue.prototype.setWorkerName = function(){
     return this.client.client('setname', this.clientName());
-  },
-  getWorkers: function(){
+  };
+
+  Queue.prototype.getWorkers = function(){
     var _this = this;
     return this.client.client('list').then(function(clients){
       return _this.parseClientList(clients);
     });
-  },
-  base64Name: function(){
+  };
+
+  Queue.prototype.base64Name = function(){
     return (new Buffer(this.name)).toString('base64');
-  },
-  clientName: function(){
+  };
+
+  Queue.prototype.clientName = function(){
     return this.keyPrefix + ':' + this.base64Name();
-  },
-  parseClientList: function(list){
+  };
+
+  Queue.prototype.parseClientList = function(list){
     var _this = this;
     var lines = list.split('\n');
     var clients = [];
@@ -36,6 +42,6 @@ module.exports = {
     });
 
     return clients;
-  }
+  };
 };
